requirements.txt
mcp>=1.0.0
pyobjc-framework-Quartz>=10.0
pyobjc-core>=10.0 